added make paper/id ssr
